Refactor linter/check module
