+++ clomonitor-core/src/linter/check/data/licenses.bin.zstd
