+++ clomonitor-core/src/linter/check/mod.rs
@@ -1,0 +1,4 @@
+pub(crate) mod content;
+pub(crate) mod github;
+pub(crate) mod license;
+pub(crate) mod path;
