+++ clomonitor-core/src/linter/check/testdata/LICENSE
