+++ clomonitor-core/src/linter/metadata.rs
@@ -62,7 +62,7 @@ mod tests {
     #[test]
     fn metadata_from_path_invalid_metadata_file() {
         assert!(matches!(
-            Metadata::from(Path::new(TESTDATA_PATH).join("LICENSE")),
+            Metadata::from(Path::new(TESTDATA_PATH).join("invalid")),
             Err(_)
         ));
     }
