+++ clomonitor-core/src/linter/mod.rs
@@ -5,7 +5,6 @@ use std::path::Path;
 use std::str::FromStr;
 
 mod check;
-mod github;
 mod metadata;
 mod patterns;
 pub mod primary;
