+++ clomonitor-core/src/linter/testdata/invalid
@@ -1,0 +1,1 @@
+invalid { metadata
